# spec/api-contract.yaml
openapi: 3.1.0
info:
  title: TTL-Aware BFF API
  version: 1.0.0
  summary: TTL-aware intelligent data-source routing Backend-for-Frontend.
  description: |
    Read-only aggregation tier over two heterogeneous sources:

      * **ODS** — Operational Data Source (gRPC, low latency, current state)
      * **EDS** — Execution Data Source (REST, higher latency, workflow / history / audit)

    Every request is classified, routed through an ordered 11-rule policy chain
    driven by configurable TTLs, source health, required fields and consistency
    requirements, mapped into a single canonical domain model, resolved by an
    explicit source-precedence policy, and returned inside a stable envelope
    carrying freshness and provenance metadata.

    Normative requirements: `spec/requirements.md`.
    Routing semantics: `spec/routing-policy.md`.
    Error semantics: `spec/error-model.md`.

    ## Response envelope

    Every 2xx body is `{ "data": ..., "meta": ... }` (REQ-API-004). `meta` always
    carries the routing decision, the stable routing rule id, the freshness
    verdict, per-field-group provenance, and structured warnings.

    ## Degradation

    `meta.degraded` means the data is old or came from a fallback source — both
    the routing-time rung (`health.primary_unavailable`) and the call-time rung
    (`fallback.primary_failed`) set it, each with a `SOURCE_UNAVAILABLE` warning.
    `meta.partial` means at least one source the request wanted did not answer.
...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
  /resources/{resourceId}:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
      - $ref: '#/components/parameters/CorrelationId'
      - $ref: '#/components/parameters/TenantIdHeader'
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/TraceParent'
      - $ref: '#/components/parameters/Consistency'
      - $ref: '#/components/parameters/Fields'
    get:
      tags: [resources]
      operationId: getResource
      summary: Read a resource
      description: |
        Classification `resource_read`. Preferred source OPERATIONAL with
        `ttl: 30s`, `cache_ttl: 5s`, `fallback: execution`, `allow_stale: true`,
        `max_stale: 300s`.

        Typical routing: `ttl.operational.fresh` → OPERATIONAL. When the
        operational record is stale and the EDS can supply the required field
        groups, `ttl.operational.stale` falls back to EXECUTION and the response
        is partial (operational-only groups such as `configuration` and
        `topology` are absent).
      x-requirement:
        - REQ-API-001
        - REQ-API-004
        - REQ-CLS-001
        - REQ-RT-001
        - REQ-TTL-002
        - REQ-EDGE-001
        - REQ-EDGE-002
        - REQ-EDGE-003
      x-request-type: resource_read
      responses:
        '200':
          $ref: '#/components/responses/ResourceOk'
        '206':
          $ref: '#/components/responses/ResourcePartial'
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
        '413': { $ref: '#/components/responses/RequestTooLarge' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/Internal' }
        '502': { $ref: '#/components/responses/UpstreamInvalid' }
        '503': { $ref: '#/components/responses/NoSourceAvailable' }
        '504': { $ref: '#/components/responses/UpstreamTimeout' }
  /resources/{resourceId}/status:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
      - $ref: '#/components/parameters/CorrelationId'
      - $ref: '#/components/parameters/TenantIdHeader'
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/TraceParent'
      - $ref: '#/components/parameters/Consistency'
      - $ref: '#/components/parameters/Fields'
    get:
      tags: [resources]
      operationId: getResourceStatus
      summary: Read a resource's status only
      description: |
        Classification `resource_status`. Preferred source OPERATIONAL with
        `ttl: 10s`, `cache_ttl: 3s`, `fallback: execution`, `allow_stale: true`,
        `max_stale: 120s`. Served by the narrow ODS RPC `GetResourceState`.

        `?consistency=strong` raises the level: the cache **read** is bypassed
        (`meta.cache.hit` is then always `false`), stale-serve is forbidden, and
        rule `consistency.strong_requires_operational` fires, reading the
        preferred source live with no TTL evaluation. There is no
        `Cache-Control: no-cache` handling on the request path.

        The required-field set for `resource_status` is `status` **alone**.
        `subState` is `omitempty` decoration that no caller can depend on, and
        listing it as required would make every answer from a source that does
        not model it look incomplete — which, since an under-covered answer is
        now a `206`, would turn a perfectly good execution-sourced status into a
        partial response. Because `status` lists the execution source as a
        supplier, a fallback answer for this endpoint stays `200`.

        When the operational record declares an `in_flight_execution_ref` and
        `routing.defaults.resolve_in_flight_execution` is on, this
        operational-only read makes one extra execution-source call, bounded by
        `in_flight_lookup_timeout` (300ms), so `/status` and `/details` cannot
        report different statuses for the same resource mid-workflow. It is
        best-effort: a failure leaves the operational answer standing. Because
        that call contributes a source, `meta.routingDecision` reads `BOTH` and
        `meta.sources` names both — the routing *decision* was still
        operational-only.
      x-requirement:
        - REQ-API-001
        - REQ-CLS-004
        - REQ-RT-001
        - REQ-TTL-003
        - REQ-CACHE-006
        - REQ-EDGE-001
        - REQ-EDGE-011
      x-request-type: resource_status
      responses:
        '200':
          description: |
            Status resolved. `data` is the narrow status projection
            (`StatusView`) rather than a full `Resource`; the groups it omits are
            omitted by design and do not set `partial` (REQ-EDGE-006).
          headers:
            X-Correlation-ID: { $ref: '#/components/headers/XCorrelationID' }
            X-BFF-Source: { $ref: '#/components/headers/XBFFSource' }
            X-BFF-Freshness: { $ref: '#/components/headers/XBFFFreshness' }
            X-BFF-Degraded: { $ref: '#/components/headers/XBFFDegraded' }
            Cache-Control: { $ref: '#/components/headers/CacheControlOut' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StatusEnvelope' }
              examples:
                fresh:
                  $ref: '#/components/examples/StatusFresh'
                staleDegraded:
                  $ref: '#/components/examples/StatusStaleDegraded'
                cacheStale:
                  $ref: '#/components/examples/StatusCacheStaleServed'
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/Internal' }
        '502': { $ref: '#/components/responses/UpstreamInvalid' }
        '503': { $ref: '#/components/responses/NoSourceAvailable' }
        '504': { $ref: '#/components/responses/UpstreamTimeout' }
  /resources/{resourceId}/configuration:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
      - $ref: '#/components/parameters/CorrelationId'
      - $ref: '#/components/parameters/TenantIdHeader'
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/TraceParent'
      - $ref: '#/components/parameters/Consistency'
      - $ref: '#/components/parameters/Fields'
    get:
      tags: [resources]
      operationId: getResourceConfiguration
      summary: Read a resource's effective configuration
      description: |
        Classification `resource_configuration`. Preferred source OPERATIONAL
        with `ttl: 30s`, `cache_ttl: 15s`, **`fallback: none`**,
        `allow_stale: true`, `max_stale: 300s`.

        `fallback: none` is structural, not a policy preference: the EDS cannot
        supply configuration, so rule `health.primary_unavailable` cannot
        redirect and the call-time `fallback.primary_failed` has nowhere to go
        either. When the ODS is unavailable the degradation ladder is
        cache → stale cache (≤ `max_stale`) → `503 NO_SOURCE_AVAILABLE`.

        **The rule id this endpoint reports is a TTL rule, not
        `fields.operational_only`.** `configuration` is exclusive to the
        operational source, so rule 4 fires — but it *pins* the source, clears
        the configured fallback and falls through to rules 8/9/10 rather than
        terminating. Terminating there would skip the `max_stale` ceiling for the
        one request type that has nowhere else to go, and that ceiling is a
        safety property rather than an optimisation. Rule 4 emits its own id only
        when the operational source is unavailable, and then with
        `routingDecision: NONE`.

        `data` is the narrow configuration projection (`ConfigurationView`),
        not a full `Resource`.

        Configuration keys matching configured secret patterns are redacted
        according to the caller's permissions (REQ-SEC-011).
      x-requirement:
        - REQ-API-001
        - REQ-RT-007
        - REQ-TTL-008
        - REQ-SEC-011
        - REQ-EDGE-003
      x-request-type: resource_configuration
      responses:
        '200':
          description: Configuration resolved from the operational source or from cache.
          headers:
            X-Correlation-ID: { $ref: '#/components/headers/XCorrelationID' }
            X-BFF-Source: { $ref: '#/components/headers/XBFFSource' }
            X-BFF-Freshness: { $ref: '#/components/headers/XBFFFreshness' }
            X-BFF-Degraded: { $ref: '#/components/headers/XBFFDegraded' }
            Cache-Control: { $ref: '#/components/headers/CacheControlOut' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConfigurationEnvelope' }
              examples:
                fresh: { $ref: '#/components/examples/ConfigurationFresh' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/Internal' }
        '502': { $ref: '#/components/responses/UpstreamInvalid' }
        '503': { $ref: '#/components/responses/NoSourceAvailable' }
        '504': { $ref: '#/components/responses/UpstreamTimeout' }
  /resources/{resourceId}/executions:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/IncludeAudit'
      - $ref: '#/components/parameters/Consistency'
      - $ref: '#/components/parameters/CorrelationId'
      - $ref: '#/components/parameters/TenantIdHeader'
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/TraceParent'
    get:
      tags: [executions]
      operationId: listResourceExecutions
      summary: List a resource's execution history
      description: |
        Classification `execution_history`. Preferred source EXECUTION with
        `ttl: 0s` (**always live**), `cache_ttl: 0s`, `fallback: none`,
        `allow_stale: false`.

        `ttl: 0` means the freshness verdict is `STALE` by construction, so the
        fresh branch is unreachable and no cached or stale answer is ever served
        (REQ-TTL-010). Rule `fields.execution_only` fires because
        `executionHistory` can only be supplied by the EDS; no operational
        substitution is permitted, so EDS unavailability is a `503` rather than a
        degraded answer (REQ-EDGE-004).

        An existing resource with no executions returns `200` with
        `data.items: []` and `data.total: 0` — never `404` (REQ-EDGE-019). No
        warning is attached: an empty collection is an answer, and there is no
        `EMPTY_RESULT` code in the canonical warning vocabulary.

        Because `ttl: 0s`, `meta.freshness` reports `state: UNKNOWN` with
        `ttlSeconds: 0`. Execution data is a record of what happened rather than
        a cache of current state, so claiming `FRESH` would imply a guarantee the
        source does not make.
      x-requirement:
        - REQ-API-001
        - REQ-API-014
        - REQ-RT-001
        - REQ-TTL-010
        - REQ-EDGE-004
        - REQ-EDGE-019
      x-request-type: execution_history
      responses:
        '200':
          description: Execution history, newest first by `startedAt`.
          headers:
            X-Correlation-ID: { $ref: '#/components/headers/XCorrelationID' }
            X-BFF-Source: { $ref: '#/components/headers/XBFFSource' }
            X-BFF-Freshness: { $ref: '#/components/headers/XBFFFreshness' }
            X-BFF-Degraded: { $ref: '#/components/headers/XBFFDegraded' }
            Cache-Control: { $ref: '#/components/headers/CacheControlOut' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExecutionListEnvelope' }
              examples:
                page: { $ref: '#/components/examples/ExecutionHistoryPage' }
                empty: { $ref: '#/components/examples/ExecutionHistoryEmpty' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/Internal' }
        '502': { $ref: '#/components/responses/UpstreamInvalid' }
        '503': { $ref: '#/components/responses/NoSourceAvailable' }
        '504': { $ref: '#/components/responses/UpstreamTimeout' }
  /resources/{resourceId}/executions/{executionId}:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
      - $ref: '#/components/parameters/ExecutionId'
      - $ref: '#/components/parameters/IncludeAudit'
      - $ref: '#/components/parameters/Consistency'
      - $ref: '#/components/parameters/CorrelationId'
      - $ref: '#/components/parameters/TenantIdHeader'
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/TraceParent'
    get:
      tags: [executions]
      operationId: getResourceExecution
      summary: Read a single execution
      description: |
        Classification `execution_status`. Preferred source EXECUTION with
        `ttl: 5s`, `cache_ttl: 2s`, `fallback: none`, **`allow_stale: false`**.

        Stale-serve is disabled because a stale execution status is actively
        misleading — a caller may believe a run has finished when it is still
        applying changes. The execution must belong to the requested resource;
        a mismatch is `404`, not a cross-resource read.

        `allow_stale: false` is expressed as `consistency: strong`, and that has
        a second consequence worth stating: a strongly-consistent request
        **bypasses the cache read**. So despite the configured `cache_ttl: 2s`
        this endpoint is never *answered* from cache — `meta.cache.hit` is always
        `false` — although a successful response is still written back through
        `Manager.Store` for callers at weaker levels (REQ-CACHE-006).
      x-requirement:
        - REQ-API-001
        - REQ-CLS-004
        - REQ-TTL-008
        - REQ-EDGE-004
        - REQ-EDGE-007
        - REQ-EDGE-015
      x-request-type: execution_status
      responses:
        '200':
          description: The execution record.
          headers:
            X-Correlation-ID: { $ref: '#/components/headers/XCorrelationID' }
            X-BFF-Source: { $ref: '#/components/headers/XBFFSource' }
            X-BFF-Freshness: { $ref: '#/components/headers/XBFFFreshness' }
            X-BFF-Degraded: { $ref: '#/components/headers/XBFFDegraded' }
            Cache-Control: { $ref: '#/components/headers/CacheControlOut' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExecutionEnvelope' }
              examples:
                running: { $ref: '#/components/examples/ExecutionRunning' }
                failed: { $ref: '#/components/examples/ExecutionFailed' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/Internal' }
        '502': { $ref: '#/components/responses/UpstreamInvalid' }
        '503': { $ref: '#/components/responses/NoSourceAvailable' }
        '504': { $ref: '#/components/responses/UpstreamTimeout' }
  /resources/{resourceId}/details:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
      - $ref: '#/components/parameters/CorrelationId'
      - $ref: '#/components/parameters/TenantIdHeader'
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/TraceParent'
      - $ref: '#/components/parameters/Consistency'
      - $ref: '#/components/parameters/Fields'
    get:
      tags: [resources]
      operationId: getResourceDetails
      summary: Read a resource with execution context (fan-out)
      description: |
        Classification `resource_details`. Preferred source BOTH with `ttl: 30s`,
        `cache_ttl: 5s`, `fallback: operational`, `allow_stale: true`,
        `max_stale: 300s`, `required_sources: {operational: true, execution: false}`.

        Rule `fields.span_both` fires because the required field set contains
        both operational-only groups (`configuration`, `topology`) and
        execution-only groups (`latestExecution`, `lastOperation`). The two
        source calls run concurrently with independent deadlines; total latency
        is the *maximum* of the two branches, never their sum (REQ-PERF-006).

        The execution branch is **optional**: its failure sets `partial: true`
        and attaches a warning, and never cancels the operational branch
        (REQ-AGG-002). The operational branch is **required**: its failure aborts
        the fan-out and sends the request down the degradation ladder.

        A partial answer is always `206`, whichever side is missing —
        `response.Writer.OK` derives the status from `meta.partial` alone, so
        there is no `200 partial: true` case. A degraded-but-complete answer
        stays `200` with `meta.degraded: true`, because the data is whole and
        only older than intended (REQ-ERR-009).

        `data` embeds the resource's own fields at the top level alongside
        `latestExecution` and `executionHistory`; there is no `data.resource`
        wrapper.

        When the latest execution is in progress, the `status` and `subState`
        fields are taken from the EXECUTION source regardless of base precedence
        (REQ-PREC-003), and `meta.provenance` reflects that.
      x-requirement:
        - REQ-API-001
        - REQ-API-006
        - REQ-API-007
        - REQ-AGG-001
        - REQ-AGG-004
        - REQ-PREC-001
        - REQ-PREC-003
        - REQ-EDGE-008
        - REQ-EDGE-015
        - REQ-EDGE-018
      x-request-type: resource_details
      responses:
        '200':
          description: |
            Details assembled from both sources. `degraded: true` when the data
            is stale or came from a fallback source. A missing source makes the
            response `206`, never a `200` with `partial: true`.
          headers:
            X-Correlation-ID: { $ref: '#/components/headers/XCorrelationID' }
            X-BFF-Source: { $ref: '#/components/headers/XBFFSource' }
            X-BFF-Freshness: { $ref: '#/components/headers/XBFFFreshness' }
            X-BFF-Degraded: { $ref: '#/components/headers/XBFFDegraded' }
            Cache-Control: { $ref: '#/components/headers/CacheControlOut' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResourceDetailsEnvelope' }
              examples:
                complete: { $ref: '#/components/examples/DetailsComplete' }
                partialExecutionTimeout: { $ref: '#/components/examples/DetailsPartial' }
                runningOverride: { $ref: '#/components/examples/DetailsRunningOverride' }
        '206':
          description: |
            At least one source the request wanted did not answer — normally the
            optional execution branch. Present fields are returned; absent ones
            are omitted, never null-filled.
          headers:
            X-Correlation-ID: { $ref: '#/components/headers/XCorrelationID' }
            X-BFF-Source: { $ref: '#/components/headers/XBFFSource' }
            X-BFF-Freshness: { $ref: '#/components/headers/XBFFFreshness' }
            X-BFF-Degraded: { $ref: '#/components/headers/XBFFDegraded' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResourceDetailsEnvelope' }
              examples:
                executionUnavailable: { $ref: '#/components/examples/DetailsExecutionUnavailable' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/Internal' }
        '502': { $ref: '#/components/responses/UpstreamInvalid' }
        '503': { $ref: '#/components/responses/NoSourceAvailable' }
        '504': { $ref: '#/components/responses/UpstreamTimeout' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Required on every `/api/v1/*` route (REQ-SEC-001).

        Validation (REQ-SEC-002): signature against the JWKS at
        `security.jwt.jwks_url` (or HS256 from `security.jwt.hs256_secret_env`
        in development only), `iss`, `aud`, `exp`, `nbf`, `iat`, with
        `security.jwt.leeway` (default 30s) applied symmetrically. `alg` is
        allow-listed; `none` and algorithm-confusion attempts are rejected. All
        `security.jwt.required_claims` must be present.

        The tenant is resolved from the token claim (`tenant_id` by default) and
        is authoritative. When `X-Tenant-ID` is also supplied it must match the
        claim, otherwise `403 TENANT_MISMATCH` (REQ-SEC-004).

        Authorization is by role → permission mapping per endpoint
        (REQ-SEC-005); see `spec/security.md`.
  parameters:
    ResourceId:
      name: resourceId
      in: path
      required: true
      description: |
        Tenant-scoped resource identifier, the join key between ODS and EDS.
        Validated before anything else touches it: letters, digits and
        `- _ . :` only, at most 128 characters. Anything else is `400`, refused
        before a data source is contacted.
      schema:
        type: string
        pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$'
        minLength: 1
        maxLength: 128
      example: res-1042
...
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      description: |
        Client-supplied correlation id, echoed in the response header and in
        `meta.correlationId`. Generated as a UUIDv4 when absent or malformed
        (REQ-API-010). Propagated to both sources.
      schema:
        type: string
        pattern: '^[A-Za-z0-9._-]{1,128}$'
      example: 2f9c1a7e-3b64-4c2a-9d15-7a0f4b8e1c33
    TenantIdHeader:
      name: X-Tenant-ID
      in: header
      required: false
      description: |
        Optional tenant assertion. When present it MUST equal the JWT tenant
        claim, otherwise `403 TENANT_MISMATCH` (REQ-SEC-004).

        In a `local` or `test` environment running with
        `security.allow_insecure_no_auth: true`, which configuration validation
        permits nowhere else, this header *is* the tenant and no token is needed;
        it defaults to `local` when absent. That is how the compose stack is
        driven.
      schema:
        type: string
        pattern: '^[a-z0-9][a-z0-9-]{0,63}$'
      example: acme
...